Remove fish shell from dotfiles
Decided to go back to just plain ol' Bash instead of using Fish.
Trying to reduce my tooling, realized that the only
reason I was using Fish was because of the nice terminal setup, which
I could achieve within bash, as well as autocomplete, which I'm not sure is
worth it.

If in the future I want to come back I'll simply revert changes.